requirements.txt
# Medical Record Inspector - Requirements
# Core dependencies for the Medical Record Inspector project.

## Data Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0

## Machine Learning & NLP
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0

## Data Science
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0

## API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

## Configuration & Logging
PyYAML>=6.0.1
colorlog>=6.7.0

## Testing
pytest>=7.4.0
pytest-cov>=4.1.0
